+++ CHANGELOG.md
@@ -11,9 +11,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - New listener for UpdraftPlus Backup/Restore plugin.
 - New listener for iThemes Security plugin.
 - Detection of environment type changes (feature introduced in WordPress 5.5).
+- The Syslog handler can now send extended timestamps (RFC5424)
 ### Changed
 - The positions of PerfOps menus are pushed lower to avoid collision with other plugins. (thanks to [Loïc Antignac](https://github.com/webaxones)).
 - The selector for WordPress events logs is now sorted: running first, paused after (thanks to [Loïc Antignac](https://github.com/webaxones)).
+### Fixed
+- Some typos in new logger UI.
 
 ## [1.13.0] - 2020-07-20
 ### Added